requirements.txt
python-hcl2
PyYAML
pytest
# pytest-mock
# mock